Simplify release (#268)
* Simplify release
* Fix typo
* Add docker release. Update build scripts

+++ docker/build-dev.sh
@@ -1,13 +1,29 @@
 #!/usr/bin/env bash
+# Build a dev image for specific framework. Optionally `all` frameworks can be built.
+# Ex: `bash docker/build-dev.sh pytorch`
+
+if [ "$#" -ne 1 ]; then
+    echo "Please pass a single argument to specify which framework to build. Must be either \`tf1\`, \`tf2\` or \`pytorch\` or \`all\`"
+    exit 1
+fi
+
+if [[ "$1" != "pytorch" && "$1" != "tf1" && "$1" != "tf2" && "$1" != "all" ]]; then
+    echo "Framework argument must be either \`tf1\`, \`tf2\` or \`pytorch\` or \`all\`"
+    exit 1
+fi
+
+# Parse Version
 version=$(python -m armory --version)
 
-docker build --force-rm --file docker/Dockerfile --target armory -t twosixarmory/armory:${version} .
-
-docker build --force-rm --file docker/tf1/Dockerfile --build-arg armory_version=${version} --target armory-tf1-base -t twosixarmory/tf1-base:${version} .
-docker build --force-rm --file docker/tf1-dev/Dockerfile --build-arg armory_version=${version} --target armory-tf1-dev -t twosixarmory/tf1:${version} .
-
-docker build --force-rm --file docker/tf2/Dockerfile --build-arg armory_version=${version} --target armory-tf2-base -t twosixarmory/tf2-base:${version} .
-docker build --force-rm --file docker/tf2-dev/Dockerfile --build-arg armory_version=${version} --target armory-tf2-dev -t twosixarmory/tf2:${version} .
-
-docker build --force-rm --file docker/pytorch/Dockerfile --build-arg armory_version=${version} --target armory-pytorch-base -t twosixarmory/pytorch-base:${version} .
-docker build --force-rm --file docker/pytorch-dev/Dockerfile --build-arg armory_version=${version} --target armory-pytorch-dev -t twosixarmory/pytorch:${version} .
+if [[ "$1" == "all" ]]; then
+  echo "Building docker images for all frameworks..."
+  for framework in "tf1" "tf2" "pytorch"; do
+    docker build --force-rm --file docker/Dockerfile --target armory -t twosixarmory/armory:${version} .
+    docker build --force-rm --file docker/${framework}/Dockerfile --build-arg armory_version=${version} --target armory-${framework}-base -t twosixarmory/${framework}-base:${version} .
+    docker build --force-rm --file docker/${framework}-dev/Dockerfile --build-arg armory_version=${version} --target armory-${framework}-dev -t twosixarmory/${framework}:${version} .
+  done
+else
+  docker build --force-rm --file docker/Dockerfile --target armory -t twosixarmory/armory:${version} .
+  docker build --force-rm --file docker/${1}/Dockerfile --build-arg armory_version=${version} --target armory-${1}-base -t twosixarmory/${1}-base:${version} .
+  docker build --force-rm --file docker/${1}-dev/Dockerfile --build-arg armory_version=${version} --target armory-${1}-dev -t twosixarmory/${1}:${version} .
+fi
